+++ src/imio/smartweb/core/contents/sections/subscriber.py
@@ -3,6 +3,18 @@
 from Acquisition import aq_parent
 
 
+def added_section(obj, event):
+    if obj.collapsible_section:
+        obj.hide_title = False
+    reindexParent(obj, event)
+
+
+def modified_section(obj, event):
+    if obj.collapsible_section:
+        obj.hide_title = False
+    reindexParent(obj, event)
+
+
 def reindexParent(obj, event):
     parent = aq_parent(obj)
     if parent is not None: