+++ src/imio/smartweb/core/tests/test_sections.py
@@ -378,6 +378,59 @@ class TestSections(ImioSmartwebTestCase):
         view.show_section_title()
         self.assertFalse(section.hide_title)
 
+    def test_sections_collapsible(self):
+        page = api.content.create(
+            container=self.portal,
+            type="imio.smartweb.Page",
+            title="Page",
+        )
+        api.content.transition(page, "publish")
+        section_types = get_sections_types()
+        for section_type in section_types:
+            api.content.create(
+                container=page,
+                type=section_type,
+                title="Title of my {}".format(section_type),
+            )
+        video_section = getattr(page, "title-of-my-imio-smartweb-sectionvideo")
+        video_section.video_url = "https://www.youtube.com/watch?v=_dOAthafoGQ"
+        gallery_section = getattr(page, "title-of-my-imio-smartweb-sectiongallery")
+        api.content.create(
+            container=gallery_section,
+            type="Image",
+            title="Image",
+        )
+        files_section = getattr(page, "title-of-my-imio-smartweb-sectionfiles")
+        file_obj = api.content.create(
+            container=files_section,
+            type="File",
+            title="My file",
+        )
+        file_obj.file = NamedBlobFile(data="file data", filename="file.txt")
+        links_section = getattr(page, "title-of-my-imio-smartweb-sectionlinks")
+        api.content.create(
+            container=links_section,
+            type="imio.smartweb.BlockLink",
+            title="My link",
+        )
+        selections_section = getattr(
+            page, "title-of-my-imio-smartweb-sectionselections"
+        )
+        intids = getUtility(IIntIds)
+        selections_section.selected_items = [
+            RelationValue(intids.getId(self.page)),
+        ]
+
+        view = queryMultiAdapter((page, self.request), name="full_view")()
+        collapsable_elems = view.count("collapse")
+        for section_id in page.objectIds():
+            section = getattr(page, section_id)
+            section.collapsible_section = True
+            modified(section)
+            self.assertFalse(section.hide_title)
+        view = queryMultiAdapter((page, self.request), name="full_view")()
+        self.assertEqual(view.count("collapse"), collapsable_elems + 17)
+
     def test_background_style(self):
         section = api.content.create(
             container=self.page,