+++ src/imio/smartweb/core/tests/test_text.py
@@ -85,7 +85,7 @@ class TestText(ImioSmartwebTestCase):
         section.alignment = "left"
         section.image_size = "large"
         view = getMultiAdapter((self.page, self.request), name="full_view")
-        self.assertIn('<div class="body-section figure-left no-image">', view())
+        self.assertIn('<div class="body-section figure-left  no-image"', view())
         self.assertNotIn("<figure", view())
         self.assertNotIn("@@images/image/large", view())
         self.assertNotIn("figcaption", view())
@@ -94,7 +94,7 @@ class TestText(ImioSmartwebTestCase):
             "ploneLeadImage", filename=get_leadimage_filename()
         )
         view = getMultiAdapter((self.page, self.request), name="full_view")
-        self.assertIn('<div class="body-section figure-left">', view())
+        self.assertIn('<div class="body-section figure-left "', view())
         self.assertIn("figure", view())
         self.assertIn("figure-left", view())
         self.assertIn("@@images/image/large", view())
@@ -103,7 +103,7 @@ class TestText(ImioSmartwebTestCase):
         section.image_caption = "Kamoulox"
         view = getMultiAdapter((self.page, self.request), name="full_view")
         # Assert section text has lead image
-        self.assertIn('<div class="body-section figure-left">', view())
+        self.assertIn('<div class="body-section figure-left "', view())
         self.assertIn("figure", view())
         self.assertIn("figure-left", view())
         self.assertIn("figcaption", view())